change locale container

+++ src/Domain/Services/ITaxaService.cs
@@ -2,6 +2,6 @@ namespace Domain.Services
 {
   public interface ITaxaService
   {
-    decimal GetTaxa();
+    double GetTaxa();
   }
 }
