+++ src/Service/Services/Taxa.cs
@@ -4,9 +4,9 @@ namespace Services
 {
   public class Taxa : ITaxaService
   {
-    public decimal GetTaxa()
+    public double GetTaxa()
     {
-      return 0.01M;
+      return 0.01;
     }
   }
 }
